Renamed expiration_monitor to secret_expiration_monitor (#7)
